pyproject.toml
[project]
name = "sam2-mask-annotation"
version = "0.1.0"
description = "Browser-based COCO mask correction tool powered by SAM2."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "numpy>=1.24.0",
  "pillow>=10.0.0",
  "pycocotools>=2.0.8",
  "pydantic>=2.7.0",
  "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]
sam2 = [
  "sam2>=1.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
